Refactor trace store import bundle helpers
